Syntax for easy mounting of services within a .war deployment.
Created an examples-war project, from which a simple `package` command
builds a portable war for servlet container deployments.  We didn't
want http4s-jetty or http4s-tomcat in this project, which caused a
split of the examples project.

As a result of this refactoring, it's now possible to conveniently start
and test any of the backends with a command like so:

   sbt examples-blaze/reStart shell

Or, for a container deployment:

   sbt examples-war/container:start shell

Closes #139, #140.