Use the gatling-it configuration
It doesn't make a lot of sense to test the demo.http4s.org site as it won't be using
the http4s master branch. This PR allows that to be done explicitly,
and can be changed once we start spinning up a local server to test
against.

Tests can be explicitly run in the load-test project with gatling-it:test