Allow PHP-style [] in query strings
PHP encodes arrays in query strings using square brackets. This goes
against the spec, but PHP ain't got no time for your spec poindexter.

This is a minimal change that only allows square brackets in keys and
values. Arguably a more permissive parser (allowing any non-separator or
terminator character in keys and values) would be better for the wild
and crazy world we call the web.