Add `defaultCharset` to the decodeString function
See issue #279 for motivation.

Many types of messages are encoded as text, and some have a default charset. In situations like this, the Content-Type header may not acutally specify the charset. See UTF-8 for an example. https://tools.ietf.org/html/rfc7159

The commit allows the specification of a default charset for the `EntityDecoder.decodeString` function for situations like these.