Explicitly specify the UTF-8 charset in json encoders
This PR does not address any known bugs. The current default charset for String encoding is UTF-8. The default String encoder is used for json encoding and thus get the correct charset, UTF-8.

This PR just makes the specification of UTF-8 explicit for documentation purposes and as insurance against any unlikely change of the default in the future.