Make convenient ETag header construct function
Problem: PR #446 (commit 0ce2754eb735d2acb94e2c45758392335a448f99) added the
type `EntityTag` to fix the malformed ETag and `If-Modified-Since` headers. In
the process construction of the `ETag` header became more difficult due to it
now containing an inner object instead of a simple `String`.

Fix: Add a simple apply function to the object to give the same behavior we had
before.

This PR should be binary backwards compatible with 0.11.0: it only modifies the
internal of the `ETag` header parser and introduces one new method to the `ETag`
object.