Extend timeout and wait durations in test to help travis builds
Problem: The Travis-ci builds fail relatively often (maybe 1/3 of the time?)
due to timeouts not happening when they should. This is likely caused by the
short timeouts and delays set in the unit tests combined with the heavy load
on the server while running tests concurrently.

Solution: Extend the durations so that threads are more likely to restart in
within their specified duration.

Bonus: make the default max entity size in the `EntityLimiter` easier to
read.