Unseal the EntityDecoder trait
Problem: The EntityDecoder trait is sealed, making it impossible to create EntityDecoder's that can access the message body.
Solution: Unseal it. The only reason to keep it sealed would be to allow totality checking with pattern matching but considering the problem EntityDecoder is intented to solve that seems like an highly improbably use case.