Replace duplicate test and bring test behavior more in line with its title
closes #484

Problem: Spec "Handle routes that ignores the body for non-chunked" didn't actually ignore its body: it echoed it. Test "Handle routes that consumes the full request body for non-chunked" echoed the body, it did not consume it.

Solution: One test was repurposed and the other renamed to be more clear about what it does.