Remove the 'clear excess bytes from the wire' feature from blaze
The behavior on master is to read all the bytes off the line until completion. That could be a real bummer for a big POST body that we don't care about. This commit removes that and simply hangs up after servicing a request if there are body bytes from the client remaining on the wire.