Fix infinite loop
When the client is asked to make a request that cannot be opened (say, invalid Uri) it will enter an infinite loop because the failure path creates a new connection with the same key, simply failing again, etc...

This commit changes that behavior by simply discarding the connection at that point and handing the origin that failure.

This commit also changes behavior: previously if the pool received a closed connection it would open a new one to the same endpoint and place it in the idle queue. This commit doesn't do that though it would be trivial to add.

This commit also changes a strange behavior on returning a connection. Previously, if there were waiting connections but they didn't match this endpoint the connection would simply be returned to the idlequeue. This causes a potential deadlock where you saturate the open connections at one endpoint, change endpoint, but the connections from the first endpoint sit in the idle queue while the waiting queue is filled with requests to the second endpoint. This commit fixes that, killing the
connection if there are pending connections that need a different endpoint.