Fix 'Stage has request in progress' leak.
If the parser is not complete when we return the body, we rely on a cleanup
hook of the body to either reset or shutdown the connection.  If this body is
not run to completion, neither happens.  This leaves the client in the
`Running` state.

Meanwhile, the pooled connection manager happily recycles any connection that
is not closed.  Hilarity ensues.

This exposes a check that the connection is recyleable, which in the blaze case
is true only when the connection has been able to reset itself.