Remove body from request.toString and response.toString.
The body of a scalaz-stream does not tend to have a very
legible toString.  Ways of making it legible often require
running the body, which should only be done once.

A process of pure Emits of character data can be turned into
a String.  We can do that, for the subset of messages that
are pure emits.  If we *don't* do that, we're still
leaking the contents of the body, but not in a format that
can be sanitized by a reasonable log filter.

We can't anticipate all the things a user wants to log or not,
but we should at least make it obvious to the user what is
being logged.