Reject invalid Uris
Fixes #547

Problem: The Uri parser doesn't demand the entire input satisfy the Uri criteria causing invalid uri's to truncate to "" which appears to be a valid Uri (at least by the parser rules).

Solution: This is solved by adding the `EOI` rule to the end of the parsers requiring the parser to consume all the input to be valid. Two tests that allowed the initial valid portion of the request to succeed how require a failure.