This doesn't pass tests but starts a discussion
Fixes #547

Problem: The Uri parser doesn't demand the entire input satisfy the Uri criteria causing invalid uri's to truncate to "" which appears to be a valid Uri (at least by the parser rules).

Solution: This is solved by adding the `EOI` rule to the end of the parsers requiring the parser to consume all the input to be valid.
    This breaks two tests in UriParserSpec: they allow invalid and invalid uri or query to essentially truncate to the valid portion. This seems wrong to me.