Add mima exception for `encodeHeaders`
The previous commits changed the signature of the function.
The function shouldn't be part of the public api anyway but
its in a bit of a funky package situation with respect to the
dependent blaze components so I didn't make it private yet: I'd
rather take care of that in the master branch.