Simplify {Http,}Service.apply methods, Response.fallthrough
The raison d'être for the apply methods is to make a partial function
total via fallthrough.  The overloads are unused complications.

Move the `HttpService.notFound` to `Response.fallthrough`.  This signals
intent more clearly for middlewares that asynchronously pass on a
request, such as the resource handlers.

Eliminate the response attribute.  Letting users override the
`Fallthrough[Response]` involves dangerous games with type class
incoherence.