Review feedback on Uri./ improvements
Eliminate charset argument to Uri./.  The other components assume UTF-8,
so we don't gain anything from this parameterization.  We need an
internal representation that preserves the encoding of certain
characters.  UTF-8 storage makes for efficient rendering in the dominant
case, and if we want to support other charsets, we can parameterize the
renderer.