Move the custom keys into build.sbt
We apparently can't set keys not defined in an *.sbt from the command
line, so those needed to move.

Once moved to *.sbt, the *.scala files can't see them to declare tasks,
so the binding of values to the hack in Http4sGhPages also had to move.
That's hopefully all temporary, because sbt/sbt-ghpages#17.