Order[CharsetRange] is unlawful, so weaken it to Equal
It's based entirely on the QValue.  This makes it possible to write an
`f: CharsetRange => CharsetRange` that determines the QValue by the
charset, and breaks `x eqv y ==> f(x) eqv f(y)`.

We could make it lawful by using the charset to break ties, but this has
no basis in the HTTP spec.  One part of the HTTP spec we missed was to
prefer atoms over splats.  All of this is specific to
content-negotation, and belongs in code there, not a general ordering.

This was uncovered by `Cogen` in cats-0.8.  It will be tested there.

/cc @ceedubs, whose morning I helped ruin with this.