Make the `getChallenge` method of bundled auth middleware public
As the Authentication middleware doesn't "fall through" in a stack, there is no way to allow users to authenticate with multiple different auth sources. In order to enable that kind of functionality, the existing `getChallenge` method of both `BasicAuth` and `DigestAuth` must be public so external integrations can properly invoke them.