Fix blaze-client eager connection closing
Problem

The blaze-client attempts to avoid running the potentially
effectful body of a request by sending only the prelude first,
hoping to get a EOF in return if the connection is stale.
Unfortunately, this doesn't work: we get a success even if
a peer that has sent a FIN (and we have even ACKed it).

Solution

Buffer one read, causing the OS to let us know that the
connection has gone down. This causes the stage to register
as closed as soon as that happens.