Get rid of delambdafy on 2.11
This was the cause of some pain in rho between 2.11.7 to 2.11.8,
whereupon @reactormonk and I were advised by the compiler folks that
these flags are exprimental and unmaintained in 2.11.  We'll leave
them in 0.15.x because binary compatibility worries me more than
what's left of these flags, but 0.16.x is a new beginning.