Fix logic to pick build where coverage and mima are run
The change to `$TRAVIS_BUILD_NUMBER == 1` in commit 96bb7c5e
did not work as expected.  That value is a counter that
increments for every Travis run.  3817 is a recent example.

This is an example build showing the wrong behavior:
https://travis-ci.org/http4s/http4s/jobs/207603576#L320
Travis is building with Scala 2.12.1 and scalaz 7.2.8, but only
running with `;test ;publish`.

Switched back to checking $TRAVIS_SCALA_VERSION and $SCALAZ_VERSION
because it works and is more explicit.  Looked for a Travis environment
variable for the position in the matrix, but it's not exposed directly.
Would need to use the suffix of $TRAVIS_JOB_NUMBER.

Also extract some of the shell script logic into named functions
so it's easier to follow the guts of the script.  Might make sense
to move those function into a separate file and source them from
bin/travis in a future commit.