increase timeout for client tests, like AHC, that connect to Jetty
AsyncHttpClientSpec would occasionally exceed the 10 second timeout
when running in 2 core containers on Travis.  Does not happen
consistently, so probably the result of a noisy neighbor or other
CPU or disk throttling.  10 seconds *should* be enough, and I suspect
there is performance still to be had with AsyncHttpClient, but the
sporadic CI failures are wasting developer time without any benefit.

Refs #1055