begin DRYing up library versions used by tuts
Add hugo shortcodes[1] for circe and cryptobits to reduce the
number of places version numbers are copied.

tut blocks are compiled against library versions in `build.sbt`
and `project/Http4sBuild.scala`, but the versions are also part
of surrounding markdown that is not compiled.  This copypasta
allows the compiled tuts to use version B of a library while a
reader following along at home pastes in older version A.  While
typically not a problem, it may allow the tuts, as rendered in
HTML, to break silently.

This change reduces the version copying, but does not sync the
versions in `Http4sBuild.scala` with the ones displayed in the
rendered HTML.

A good follow-up step might be to generate hugo's `config.toml`
from sbt.  Or, maybe better, export version from sbt as
environment variables and use getenv[2] in the shortcode templates.

Also updated version for `argonaut-shapeless_6.2` to latest
milestone so all the versions are up-to-date.

[1] https://gohugo.io/extras/shortcodes/
[2] http://gohugo.io/templates/functions/#getenv