Link to different branches from scaladoc source links
If the current git commit has a tag, link to the git tag of the
current version.
If the current git commit has no tag and the current branch is
master, link to the master branch.
If the current git commit has no tag and the current branch is not
master, link to the current git commit.