Make sure trailers are not null when closing stream.
grpc-java expects that status and trailers are non-null when closing
the stream:
https://github.com/grpc/grpc-java/blob/master/core/src/main/java/io/grpc/internal/AbstractServerStream.java#L128

If the exception is raised like
```scala
io.grpc.Status.INTERNAL
  .withDescription("raising error")
  .asRuntimeException())
```
then this will not actually close the stream because trailers is null.

Ddue to the signature of the stream, Status-based exceptions
are currently the only way to signal specific types of gRPC failure.