Docs: improve client doc
The documentation says that when using `.stream` you'll receive
a `Stream[F, Http1Client]`, this is false. The method returns a
`Client[F]`.

I've also added the type on the `val` below just to make it obvious
what you have at this point. To me, it was not when I first reached
into the doc.