Authentication: simplify code.
We extract an "unauthorized" function, that generates a response with the
"Unatuthorized" status with the given authentication challenge.

We edit the code of challenged, so as to avoid using the Kleisli and
OptionT combinators, which can be unclear, and to use instead a
style in which the flatmaps flow as the execution would.