CSRF: Make plain
We refactor the code of CSRF to avoid the use of the special "EitherT"
methods and combinators, in favour of the base types from the Scala
library. The code is slightly longer, but more amenable to non-cats
experts.

We extract a common code for these two methods as a private method.