AuthMiddleware / Apply: simplify code
The `apply` method from the AuthMiddleware object bulds a service out
of three elements: a function to get the auth info from a request,
the authenticated service to run if it is authenticated, and the
fallback service to run if authentication fails.

We simplify the code of this apply method by doing the following:
- We avoid the methods `local` and `mapF` of Kleisli.
- We avoid using the `Choice` instance for `Kleisli[OptionT]`.
- We avoid calling the `AuthedRequest(Kleisli)` method.

This leaves the code of the function in a direct style.