Remove EitherT from the DecodeResult.
The type of DecodeResult was defined as an EitherT, which is a wrapper
for an `F[Either[L, R]]`.
It turns out that, for most operations, the full power of the EitherT
was not being used, and only one or two operations were being used
when combining stuff.

For this reason, we drop the use of EitherT. We reduce the type
DecodeResult to be just the naked "F[Either[DecodeResult, A]".

To keep the convenience we have at certain places, we introduce a
syntax extension (an implicit class), "FEither", which replicates
some of the methods of `EitherT`, but without colliding with the
names `map` and `flatMap`.

We call it FEither because we intend to keep it very light.