Fix Bug Where The Body Of The Request Is Not Sent In Async Client
The old code used the `EmptyBodyGenerator` in the event that the a request had `_.isChunked == false` and `_.contentLength.isEmpty`. This caused valid requests created in the following manner to not have their body's sent.

```scala
Request(method = POST, Stream.emit(1.toByte))
```

This seems like a strange way to construct a `Request` value, and indeed it is, but since it is permissible by http4s it should be valid.

The same issue is present in the `JavaNetClient` and is fixed in a similar way.