Add Mechanism To Create Validated Timeout Values
The impetus for this commit is the `verifyTimeoutRelations` method in `BlazeClientBuilder`. That method provides a runtime check with a warn log that the timeout values for a client pass some sanity checks.

This commit adds a new ADT `Timeout` and a product of all timeout types `Timeouts`. The latter provides a mechanism to ensure that the timeout values constructed are always valid.