Close urlConn in StaticFile.fromURL
The URLConnection in StaticFile.fromURL is never closed when the resource is not expired.
This PR reuses the already opened URLConnection to read the response body which in effect causes fs2 to close the URLConnection once the body is fully read.
Or was the body intended to open a separate url connection? In that case the previous URLConnection should probably be closed before returning the response.