Move main code of server, client, and dsl into the `core`.
WE merge the source code of the `main` folders of the `server`,
`client`, and `dsl` modules into the `core` module.

For many applications of http4s, and even libraries built on top of
it, one is unlikely to use only the `core` module alone: instead, one
is likely to end up using either the `server` or the `client` module.
Given that these modules add no library dependencies over  `core`
itself, there is no problem to integrate them further down.

Regarding the `dsl` module, given that this `dsl` has the main spot
in the http4s tutorials, and that it is more integrated with the
`Uri` and `Request` constructors, perhaps we can integrate it
further.

The `test` code, however, because they have some special test
dependencies on other modules (above the `core`), so we need to avoid
circular dependencies.