Add Order/Hash/Show Instances For Uri And Related Types
This commit adds `Order`, `Hash`, and `Show` instances for the following types.

* `org.http4s.Uri.Authority`
* `org.http4s.Query`
* `org.http4s.Uri.Host`
* `org.http4s.Uri.RegName`
* `org.http4s.Uri`

The impetus for all of them was to support instances for `org.http4s.Uri`.

In addition, two functions were added to the package object `org.http4s.internal` scoped to be private to `http4s`, `compareField`, `reduceComparisons_`, and `reduceComparisons`.

These functions substantially reduce the amount of code required to write `Order` instances for Product types.

In order to support testing the laws for these types, several new instances of `Arbitrary` and `Cogen` were added, as not all types had instances for each of these.