Handle CompletableFuture Correctly
This commit adds a new function `fromCompletableFutureShift` which brackets the usage of a `CompletableFuture` so that if the outer effect is canceled or terminates in error, then the `CompletableFuture` is canceled or terminated as well.

This is _similar_ to the deprecated `fromCompletableFuture`, but differs in that it handles exceptional cases, not just cancellation, and properly handles the thread shifting, as `fromCompletionStage` does.

`fromCompletionStage` is used in `http4s-jdk-http-client` and under the current usage causes a resource leak if the outer effect terminates before the callback has been invoked.