Update Prometheus Client To 0.10.0 And Fix Tests
Prometheus is changing the naming scheme of some of their metrics in order to be compatible with the OpenMetrics standard. Part of this means that all counters will have `_total` or `_created` as a suffix. There is no way to create a counter without a `_total` suffix.

See: https://github.com/prometheus/client_java/pull/615