FileService: Querying /file.png/anything where /file.png is a file fails
Exception is `java.nio.file.FileSystemException: Not a directory`.
Found bot queries in Scala Algorithms logs that are leading to an HTTP 5xx error.

To avoid this, we explicitly check if what we are querying is a file, before attempting to resolve the real path; the exception is currently raised in particular where `.toRealPath` is called.

Alternative approach would be to capture the `FileSystemException` as well, but it could hide other errors like `AccessDeniedException`. Because we already do a check, the check for `NoSuchFileException` becomes redundant.