Initial Encoding Of ServerBuildable
This commit provides an initial encoding a of `ServerBuildable` a pseudo typeclass alternative to `ServerBuilder`.

The intent here is to provide a clean minimal ad-hoc polymorphic way to define an interface for things which can configure and run servers.

Prior to this, the code required to build a server varied a great deal between the different backends. `ServerBuilder` provides somewhat of a common interface, but lacks even some very basic items, e.g. there is no way to pass in an `HttpApp` to `ServerBuilder`.

This leads to confusion among users about exactly how to configure and start a specific server, and discourages trying different http backends as users have to learn the idioms of each underlying system again. To an extent, that is unavoidable. Certain aspects of configuration will _always_ be specific a given backend, however it need not be quite a bespoke as it is.

Currently `ServerBuildable` implements 1 fundamental operation for serving a server,

```scala
  /** Returns the backend as a resource.  Resource acquire waits
    * until the backend is ready to process requests.
    */
  def resource(a: A): Resource[F, Server]
```

And 3 fundamental operations for configuring a server builder,

```scala
  /** Set the http host for the server. */
  def withHttpHost(a: A)(host: String): A

  /** Set the http port for the server. */
  def withHttpPort(a: A)(port: Int): A

  /** Set the [[HttpApp]] to serve on this server. */
  def withHttpApp(a: A)(app: HttpApp[F]): A
```