Add DataTypes For Extracting Out The Non-Body Portion Of A Message
This commit adds two data types, `RequestInfo` and `ResponseInfo`, which extract out the non-body portion of a `Request[F]` and `Response[F]`.

This data types can be used in many contexts, where we explicitly _don't_ want to deal with the body.

The only field these data types omit, other than the body, is `attributes` as that would prevent an `Order` instance from being written and it is unlikely that _most_ uses of these data types will explicitly care about the `attributes`. Of course, if they do, it is trivial to create a `(Vault, RequestInfo)` tuple.