EntityEncoder: use variance of type parameters
We modify the EntityEncoder class to make it covariant on the Stream
effect type F and contravariant on the type of values A to encode.

With variance, we can remove the `F[_]` type parameters from some
entity encoder generators, and just reduce it to fs2.Pure (an alias
of Nothing for kind F[_]); as well as reduce some  `def` to `val`.