Routing on variable segments
Router and ContextRouter are enhanced with support for routing on
variable path segments through the new methods `dynamic` and `of`.
Static and dynamic routing is unified through a new type
`{Router, ContextRouter}.Routable`.
Capturing and transforming a segment of the path is implemented as a
tiny middleware `{Router, ContextRouter}.Segment`.

As there are no changes to existing methods, this should be
backwards compatible. `of` and `dynamic` are source compatible with `apply` and `define`,
with only slightly worse type-inference for static cases), and
could replace `apply` and `define` at the next point of breaking binary compatibility.

The routing is always done static, with longest match first, and then
dynamic in the order they are defined. (Same as it used to be, but
with the dynamic ones at the end)