Ember Core: avoid string allocation in method parse
Current code for parsing method always creates a `String` to cover the
region of the array. This string is an additional object which has to
copy the positions of the byte array, and then is discarded.

Instead, we can just use existing methods, and compare their names
to the bytes. This avoids the extra allocation.
On the other hand, unlike the map which could compare with few, or may
be using a Hash, here we compare with all methods listed.