Should Metrics classifierF use Generic vs String?
This is a proof of concept for whether or not the `Metrics` middleware
should use a generic type instead of a `String` for `classifierF`. When
implementing a `MetricsOps`, I found myself wanting to pass along more
information than is easily included in a string. My specific use case is
reporting metrics to datadog, and wanting to synthesize both a metric
name and a set of metric tags from the request path, e.g.,

```scala
case class Context(name: String, tags: Seq[(String, String)])
```

I was able to solve this by converting a `Request[F]` into a `Context`
and then serializing to json and deserializing within my `MetricsOps`
implementation, but I had the thought that this round trip could be
avoided if the classifier worked with a type besides just `String`.

This PR is mostly a proof-of-concept for how this could look, happy
to close this and move into an issue if that would be better for
discussion.