Throttle Server middleware:
The `takeToken` method uses the Ref.access method, which yields a
setter that yields true if it can set the value, or false if it
cannot, maybe because of a concurrent write. In the latter case,
the method has to re-try by using a recursive call to `takeToken`.

Currently, that recursion goes through emitting an Option and
flatMapping. This commit changes that to a direct call.