EntityEncoder - Avoid implicit encoders for streams.
Streams of items can be encoded in many ways. The Core defines a
simple definition, which is to append without separators the encoding
of each item into the body. This may cause some problems.
On the other hand, an enconding to Json may choose to wrap items in a
Json array, as if the stream was just a list sent in parts.

Implicits should only be used when the answer "How to implement
operation Fili for type Kili` is almost unique. Since the question
"How to encode a stream" has several answers, it is best to avoid
implicit defs, and let users of library explicitly choose.