Remove Entity Apply
Since we may want to steer code away from the streamed entities as a
default, we should deprecate the Entity.apply method in favour of a
explicit `stream` constructor...

We also rename the Entity.Default class to Entity.Streamed.

Additionally, we add a couple of constructors for strict entities,
`string` and `utf8String`.