Ember-Core: reimplement write loop without streams.
The `writeLoop` was using FS2 streams without strict need. The streams
in question were generated and immediately processed.

For this use case, we can make use of a newer Queue operation,
tryTakeN, which if given a `None` parameter just dumps the entire
queue on each iteration. The main writeLoop process can just be
written as a recursive process in `F`.