Ember-Core H2Client: split method fromSocket into procedures.
Small refactor of the code in the `fromSocket` method of the
`H2Client` class, for creating and starting up an `H2Connection`.
We split the long code of the `for` comprehension (in Resource)
into a set of auxiliary methods.

Each of the auxiliary methods operates at the level of the
"F" effect, and they are only lift to a Resource for the
purposes of background running.

Initialising the `H2Connection` object is done in a separate
method. Instead of accessing the components from the variables
in the for loop, the other auxiliary methods access them
through the `H2Connection` fields themselves.