Fixes HTTP/2 sending invalid sized frames (refers to #6844)
In the current implementation, HTTP/2 data frames of any size are sent to the client, but clients may not receive frames of sizes bigger than their max frame size settings. That causes both Firefox and CURL to drop connection before receiving the body data.

This commit solves that by inspecting the remoteSettings and looking for the acceptable frame size and them splitting larger chunks in that max frame size.