Special Case Error Handler For Unparsable Request Line
This commit adds a special case error handler to be executed when the request-line, e.g. "GET / HTTP/1.1", it not able to be parsed.

This is a different case than the more general error handler we already have. RFC 7230 states a server should return either 400 or 301. The handler is configurable, but the default is to return a 301

Closes #6931