Ember-Core: H2Connection
The last two fields represent the case in which either headers or a
push promise are made of several frames. However, it is never the
case that we have both headers and a push promise in progress.
Thus, we can factor out model from tuple of options to option
of tuple of either and list.

We add a couple auxiliary methods to clear or set progress.