Removed reference to TSec in documentation
I removed the reference to the TSec library in the documentation as it
seems like it is in the process of being deprecated. The original repo
is archived and the fork https://github.com/davenverse/tsec has this in
the readme: This software is in the process of finding its way out of
your builds. This is maintained only for compatibility and security fixes.
We hope you will work with others to find solutions to this space for the
Scala community.